add request helpers
